refactor(combat): split naval resolver types, detect, and resolve/apply (#3983)
Finish the deferred phase-3 naval file seam so detection and resolve/apply are separate libraries without changing combat semantics.
